bug fixes for .pub

+++ fileTracker/file_operations.py
@@ -7,13 +7,13 @@ def rename_folder_with_pathlib(old_name, new_name):
     try:
         path = Path(old_name)
         path.rename(new_name)
-        print(f"Folder renamed from '{old_name}' to '{new_name}' successfully.")
+        pass#(f"Folder renamed from '{old_name}' to '{new_name}' successfully.")
     except FileNotFoundError:
-        print(f"The folder '{old_name}' was not found.")
+        pass#(f"The folder '{old_name}' was not found.")
     except PermissionError:
-        print(f"You don't have permission to rename '{old_name}'.")
+        pass#(f"You don't have permission to rename '{old_name}'.")
     except Exception as e:
-        print(f"Error renaming folder: {e}")
+        pass#(f"Error renaming folder: {e}")
 
 
 def create_file_in_folder(file_path):
@@ -25,26 +25,26 @@ def create_file_in_folder(file_path):
     with open(file_path, 'w'):
         pass  # Write content to the file
 
-    print(f"File '{file_path}' created successfully!")
+    pass#(f"File '{file_path}' created successfully!")
 
 def create_folder_with_pathlib(folder_path:Path):
     path = folder_path
     try:
         path.mkdir(parents=True, exist_ok=True)
-        print(f"Folder '{folder_path}' created successfully!")
+        pass#(f"Folder '{folder_path}' created successfully!")
     except FileExistsError:
-        print(f"Folder '{folder_path}' already exists!")
+        pass#(f"Folder '{folder_path}' already exists!")
     except Exception as e:
-        print(f"An error occurred: {e}")
+        pass#(f"An error occurred: {e}")
 
 def delete_folder_with_pathlib(folder_path):
     """Deletes a folder and its contents using pathlib and shutil."""
     path = Path(folder_path)
     if path.exists() and path.is_dir():
         shutil.rmtree(folder_path)
-        print(f"Folder '{folder_path}' and its contents deleted successfully!")
+        pass#(f"Folder '{folder_path}' and its contents deleted successfully!")
     else:
-        print(f"Folder '{folder_path}' does not exist or is not a directory.")
+        pass#(f"Folder '{folder_path}' does not exist or is not a directory.")
 
 def is_folder_accessible(folder_path):
     for root, dirs, files in os.walk(folder_path):
@@ -73,17 +73,17 @@ def copy_file(source, destination):
         source (str): Path to the source file.
         destination (str): Path to the destination file or directory.
     """
-    if source.stat().st_size == destination.stat().st_size:
-        return
     try:
+        if source.stat().st_size == destination.stat().st_size:
+            return
         shutil.copy(source, destination)
-        print(f"File copied from '{source}' to '{destination}' successfully!")
+        pass#(f"File copied from '{source}' to '{destination}' successfully!")
     except FileNotFoundError:
-        print(f"Source file '{source}' does not exist!")
+        pass#(f"Source file '{source}' does not exist!")
     except PermissionError:
-        print("Permission denied! Check your access rights.")
+        pass#("Permission denied! Check your access rights.")
     except Exception as e:
-        print(f"An error occurred: {e}")
+        pass#(f"An error occurred: {e}")
 
 def copy_set_file(path,source,destination):
     base_path = Path(source)
@@ -102,13 +102,13 @@ def rename_file(old_name:Path, new_name):
             return
         
         file_path.rename(new_name)
-        print(f"File renamed from '{old_name}' to '{new_name}' successfully.")
+        pass#(f"File renamed from '{old_name}' to '{new_name}' successfully.")
     except FileNotFoundError:
-        print(f"The file '{old_name}' was not found.")
+        pass#(f"The file '{old_name}' was not found.")
     except PermissionError:
-        print(f"You don't have permission to rename '{old_name}'.")
+        pass#(f"You don't have permission to rename '{old_name}'.")
     except Exception as e:
-        print(f"Error renaming file: {e}")
+        pass#(f"Error renaming file: {e}")
 
 def rename_set_file(path, source, destination):
     base_path = Path(source)
@@ -123,16 +123,19 @@ def rename_set_file(path, source, destination):
         rename_folder_with_pathlib(destination_path/relative_previous_path, new_destination)
     else:
         rename_file(destination_path/relative_previous_path, new_destination)
-        if new_destination.stat().st_size == 0:
-            copy_file(new_name,new_destination)
+        try:
+            if new_destination.stat().st_size == 0:
+                copy_file(new_name,new_destination)
+        except FileNotFoundError:
+            pass
 
 def delete_file(file_path):
     """Deletes a file."""
     if os.path.exists(file_path):
         os.remove(file_path)
-        pass#print(f"File '{file_path}' deleted successfully!")
+        pass#(f"File '{file_path}' deleted successfully!")
     else:
-        pass#print(f"File '{file_path}' does not exist!")
+        pass#(f"File '{file_path}' does not exist!")
 
 def delete_file_set(path,source,destination):
     base_path = Path(source)
@@ -142,16 +145,22 @@ def delete_file_set(path,source,destination):
     if destination_path.is_dir():
         delete_folder_with_pathlib(destination_path)
     else:
-        delete_file(destination_path)
+        try:
+            if '.pub' in str(destination_path) and source_path.exists():
+                pass
+            else:
+                delete_file(destination_path)
+        except FileNotFoundError:
+            pass
 
 def create_file(file_path):
     """Creates a file with the given content."""
     if os.path.exists(file_path):
-        pass#print(f"File '{file_path}' already exists!")
+        pass#(f"File '{file_path}' already exists!")
     else:
         with open(file_path, 'w'):
             pass
-        pass#print(f"File '{file_path}' created successfully!")
+        pass#(f"File '{file_path}' created successfully!")
 
 def create_file_set(path,source,destination):
     base_path = Path(source)
